Adds Capistrano deploy. Updates Gemfile

+++ config/unicorn.rb
@@ -1,5 +1,18 @@
-worker_processes 3
+worker_processes 2
+
+app_root = "/var/www/nourish"
+shared = "#{app_root}/shared"
+working_directory = "#{app_root}/current"
+
+listen "#{shared}/sockets/unicorn_nourish.sock", :backlog => 1024
+
 timeout 180
+
+pid "#{shared}/pids/unicorn.pid"
+
+stderr_path "#{shared}/log/unicorn.stderr.log"
+stdout_path "#{shared}/log/unicorn.stdout.log"
+
 preload_app true
 
 GC.respond_to?(:copy_on_write_friendly=) and  GC.copy_on_write_friendly = true
@@ -8,12 +21,13 @@ before_fork do |server, worker|
   defined?(ActiveRecord::Base) and ActiveRecord::Base.connection.disconnect!
 
   # kills old children after zero downtime deploy
-  old_pid = "#{server.config[:pid]}.oldbin"
-  if old_pid != server.pid
+  old_pid = "#{shared}/pids/unicorn.pid.oldbin"
+  if File.exists?(old_pid) && server.pid != old_pid
     begin
       sig = (worker.nr + 1) >= server.worker_processes ? :QUIT : :TTOU
       Process.kill(sig, File.read(old_pid).to_i)
     rescue Errno::ENOENT, Errno::ESRCH
+      # someone else did our job for us
     end
   end
